requirements.txt
appdirs==1.4.4
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cssselect==1.2.0
distro==1.9.0
fake-useragent==0.1.11
h11==0.14.0
idna==3.4
ijson==3.1.4
importlib-metadata==6.6.0
JPype1==1.5.0
lxml==4.9.2
numpy==1.23.4
outcome==1.2.0
packaging==23.2
pandas==1.5.1
parse==1.19.0
pycparser==2.21
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
requests-html==0.10.0
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tabula-py==2.9.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
w3lib==2.0.1
websockets==10.4
wsproto==1.2.0
zipp==3.10.0
